refactor(tui): break screens.py into screens/ subpackage (#1870)
Split the 2627-line monolithic screens.py into a screens/ subpackage
with one module per logical screen group: _base (shared widgets/mixins),
login, main, workspace_detail, workspace_form, and server. The package
__init__.py re-exports all public names so existing imports are unchanged.

Closes #1870
